SDK and Demo app now support 64 bit architectures
Fixed warnings for int conversion for 64 bit builds

+++ PhotoPayCloudDemo/PhotoPayCloudDemo/View/Components/PPStyledPageControl.h
@@ -22,7 +22,7 @@ typedef enum
 
 @property (nonatomic, strong) UIColor *coreNormalColor, *coreSelectedColor;
 @property (nonatomic, strong) UIColor *strokeNormalColor, *strokeSelectedColor;
-@property (nonatomic, assign) int currentPage, numberOfPages;
+@property (nonatomic, assign) NSInteger currentPage, numberOfPages;
 @property (nonatomic, assign) BOOL hidesForSinglePage;
 @property (nonatomic, assign) PageControlStyle pageControlStyle;
 @property (nonatomic, assign) int strokeWidth, diameter, gapWidth;
